third_party/blink/web_tests/external/wpt/html/infrastructure/urls/resolving-urls/query-encoding/utf-8_include=css-expected.txt

This is a testharness.js-based test.
[FAIL] CSS <link> (windows-1252) #<id> { background-image:<url> }
  assert_equals: expected "%C3%A5" but got "%E5"
[FAIL] CSS <link> (windows-1252) #<id> { border-image-source:<url> }
  assert_equals: expected "%C3%A5" but got "%E5"
[FAIL] CSS <link> (windows-1252) #<id>::before { content:<url> }
  assert_equals: expected "%C3%A5" but got "%E5"
[FAIL] CSS <link> (windows-1252) @font-face { font-family:<id>; src:<url> } #<id> { font-family:<id> }
  assert_equals: expected "%C3%A5" but got "%E5"
[FAIL] CSS <link> (windows-1252) #<id> { display:list-item; list-style-image:<url> }
  assert_equals: expected "%C3%A5" but got "%E5"
[FAIL] CSS <link> (windows-1252) #<id> { cursor:<url>, pointer }
  assert_equals: expected "%C3%A5" but got "%E5"
Harness: the test ran to completion.
